backend_p/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
Pillow>=10.2.0
aiofiles>=23.2.0
pytest>=8.0.0
httpx>=0.27.0
